fix: sizing issue with scale UI
Fixes #18. Height of webview should be the height of the contents inside
it. Previously, scale factors increased fixed height of the container
webview, thus creating empty space (e.g. at 200% scaling, height of
webview was double the height of the inner HTML page). Now height of
webview is always the height of the contained element.